Move PolicyDefinition class to the policy-spi module
