+++ extensions/control-plane/api/data-management/policydefinition-api/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/policy/PolicyDefinitionApiExtension.java
@@ -24,7 +24,7 @@ import org.eclipse.dataspaceconnector.api.transformer.DtoTransformerRegistry;
 import org.eclipse.dataspaceconnector.spi.WebService;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.event.EventRouter;
-import org.eclipse.dataspaceconnector.spi.observe.policydefinition.PolicyDefinitionObservableImpl;
+import org.eclipse.dataspaceconnector.spi.policy.observe.PolicyDefinitionObservableImpl;
 import org.eclipse.dataspaceconnector.spi.policy.store.PolicyDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
