+++ extensions/control-plane/api/data-management/policydefinition-api/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/policy/service/PolicyDefinitionServiceImplTest.java
@@ -17,7 +17,7 @@ package org.eclipse.dataspaceconnector.api.datamanagement.policy.service;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.asset.AssetSelectorExpression;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
-import org.eclipse.dataspaceconnector.spi.observe.policydefinition.PolicyDefinitionObservable;
+import org.eclipse.dataspaceconnector.spi.policy.observe.PolicyDefinitionObservable;
 import org.eclipse.dataspaceconnector.spi.policy.PolicyDefinition;
 import org.eclipse.dataspaceconnector.spi.policy.store.PolicyDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
