+++ spi/control-plane/policy-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/observe/PolicyDefinitionObservable.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.spi.observe.policydefinition;
+package org.eclipse.dataspaceconnector.spi.policy.observe;
 
 import org.eclipse.dataspaceconnector.spi.observe.Observable;
